+++ EssentialApp/EssentialAppTests/FeedLoaderCacheDecoratorTests.swift
@@ -20,7 +20,7 @@ final class FeedLoaderCacheDecorator: FeedLoader {
     }
 }
 
-class FeedLoaderCacheDecoratorTests: XCTestCase {
+class FeedLoaderCacheDecoratorTests: XCTestCase, FeedLoaderTestCase {
     
     func test_load_deliversFeedOnLoaderSuccess() {
         let feed = uniqueFeed()
@@ -33,31 +33,5 @@ class FeedLoaderCacheDecoratorTests: XCTestCase {
         let loader = FeedLoaderStub(result: .failure(anyNSError()))
         let sut = FeedLoaderCacheDecorator(decoratee: loader)
         expect(sut, toCompleteWith: .failure(anyNSError()))
-    }
-    
-    // MARK: - Helpers
-    
-    private func expect(_ sut: FeedLoader,
-                        toCompleteWith expectedResult: FeedLoader.Result,
-                        file: StaticString = #filePath,
-                        line: UInt = #line) {
-      let exp = expectation(description: "Wait for load completion")
-      
-      sut.load { receivedResult in
-        switch (receivedResult, expectedResult) {
-        case let (.success(receivedFeed), .success(expectedFeed)):
-          XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
-          
-        case (.failure, .failure):
-          break
-          
-        default:
-          XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-        }
-        
-        exp.fulfill()
-      }
-      
-      wait(for: [exp], timeout: 1)
     }
 }
