+++ EssentialApp/EssentialAppTests/FeedLoaderWithFallbackCompositeTests.swift
@@ -9,7 +9,7 @@ import XCTest
 import EssentialFeed
 import EssentialApp
 
-class FeedLoaderWithFallbackCompositeTests: XCTestCase {
+class FeedLoaderWithFallbackCompositeTests: XCTestCase, FeedLoaderTestCase {
 
   func test_load_deliversPrimaryFeedOnPrimarySuccess() {
     let primaryFeed = uniqueFeed()
@@ -45,29 +45,5 @@ class FeedLoaderWithFallbackCompositeTests: XCTestCase {
     trackForMemoryLeaks(fallbackLoader, file: file, line: line)
     trackForMemoryLeaks(sut, file: file, line: line)
     return sut
-  }
-  
-  private func expect(_ sut: FeedLoader,
-                      toCompleteWith expectedResult: FeedLoader.Result,
-                      file: StaticString = #filePath,
-                      line: UInt = #line) {
-    let exp = expectation(description: "Wait for load completion")
-    
-    sut.load { receivedResult in
-      switch (receivedResult, expectedResult) {
-      case let (.success(receivedFeed), .success(expectedFeed)):
-        XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
-        
-      case (.failure, .failure):
-        break
-        
-      default:
-        XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-      }
-      
-      exp.fulfill()
-    }
-    
-    wait(for: [exp], timeout: 1)
   }
 }
